# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Jettster service APIs.",
        "title": "Jettster Swagger API",
        "contact": {},
        "license": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/book": {
            "get": {
                "summary": "List of Books",
                "parameters": [
                    {
                        "type": "string",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto_book.Summary"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Create Book",
                "operationId": "create-book",
                "parameters": [
                    {
                        "description": "the request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto_book.Create"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto_book.Full"
                        }
                    }
                }
            }
        },
        "/book/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "summary": "Single Book",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url params",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto_book.Full"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto_book.Create": {
            "type": "object",
            "required": [
                "name",
                "page"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                }
            }
        },
        "dto_book.Full": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "page": {
                    "type": "integer"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "dto_book.List": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "string"
                },
                "page": {
                    "type": "string"
                }
            }
        },
        "dto_book.Summary": {
            "type": "object",
            "properties": {
                "books": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto_book.Full"
                    }
                }
            }
        }
    }
}